Feat : 상품 수정 및 삭제 권한 체크

+++ src/main/java/com/mypill/domain/product/Service/ProductService.java
@@ -10,6 +10,7 @@ import com.mypill.domain.product.dto.request.ProductRequest;
 import com.mypill.domain.product.dto.response.ProductResponse;
 import com.mypill.domain.product.entity.Product;
 import com.mypill.domain.product.repository.ProductRepository;
+import com.mypill.global.rq.Rq;
 import com.mypill.global.rsData.RsData;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -28,17 +29,13 @@ public class ProductService {
     private final NutrientService nutrientService;
     private final CategoryService categoryService;
     private final MemberService memberService;
+    private final Rq rq;
 
     @Transactional
     public RsData<ProductResponse> create(ProductRequest request) {
 
-        List<Nutrient> nutrients = nutrientService.findByIdIn(request.getNutrients().stream()
-                .map(Nutrient::getId)
-                .collect(Collectors.toList()));
-
-        List<Category> categories = categoryService.findByIdIn(request.getCategories().stream()
-                .map(Category::getId)
-                .collect(Collectors.toList()));
+        List<Nutrient> nutrients = MappingNutrient(request.getNutrients());
+        List<Category> categories = MappingCategory(request.getCategories());
 
         Member seller = memberService.findById(request.getSellerId()).orElse(null);
         Product product = Product.of(request, nutrients, categories, seller);
@@ -63,39 +60,40 @@ public class ProductService {
     }
 
     @Transactional
-    public RsData<ProductResponse> update(Long productId, ProductRequest request) {
-
-        Product product = findById(productId).orElse(null);
-        if(product == null){
-            return RsData.of("F-1", "존재하지 않는 상품입니다.", ProductResponse.of(product));
-        }
-
-        List<Nutrient> nutrients = nutrientService.findByIdIn(request.getNutrients().stream()
-                .map(Nutrient::getId)
-                .collect(Collectors.toList()));
-
-        List<Category> categories = categoryService.findByIdIn(request.getCategories().stream()
-                .map(Category::getId)
-                .collect(Collectors.toList()));
-
-        product.update(request, nutrients, categories);
-
-        productRepository.save(product);
-        return RsData.of("S-1", "상품 수정이 완료되었습니다.", ProductResponse.of(product));
-    }
-
-    @Transactional
-    public RsData<ProductResponse> delete(Long productId) {
+    public RsData<Product> update(Long productId, ProductRequest request) {
 
         Product product = findById(productId).orElse(null);
         if(product == null){
             return RsData.of("F-1", "존재하지 않는 상품입니다.");
         }
 
+        if(!hasPermisson(product.getSeller().getId())){
+            return RsData.of("F-2", "수정 권한이 없습니다.");
+        }
+
+        List<Nutrient> nutrients = MappingNutrient(request.getNutrients());
+        List<Category> categories = MappingCategory(request.getCategories());
+        product.update(request, nutrients, categories);
+
+        return RsData.of("S-1", "상품 수정이 완료되었습니다.", product);
+    }
+
+    @Transactional
+    public RsData<Product> delete(Long productId) {
+
+        Product product = findById(productId).orElse(null);
+        if(product == null){
+            return RsData.of("F-1", "존재하지 않는 상품입니다.");
+        }
+
+        if(!hasPermisson(product.getSeller().getId())){
+            return RsData.of("F-2", "삭제 권한이 없습니다.");
+        }
+
         product = product.toBuilder().deleteDate(LocalDateTime.now()).build();
         productRepository.save(product);
 
-        return RsData.of("S-1", "상품 삭제가 완료되었습니다.", ProductResponse.of(product));
+        return RsData.of("S-1", "상품 삭제가 완료되었습니다.", product);
     }
 
     public Optional<Product> findById(Long productId){
@@ -105,11 +103,27 @@ public class ProductService {
     public List<Product> findAll() {
         return productRepository.findAll();
     }
+
     public List<Product> findNotDeleted() {
         return productRepository.findByDeleteDateIsNull();
     }
 
+    private  List<Nutrient> MappingNutrient(List<Nutrient> nutrients){
+        return nutrientService.findByIdIn(nutrients.stream()
+                .map(Nutrient::getId)
+                .collect(Collectors.toList()));
+    }
+    private  List<Category> MappingCategory(List<Category> categories){
+        return categoryService.findByIdIn(categories.stream()
+                .map(Category::getId)
+                .collect(Collectors.toList()));
+    }
+
     private ProductResponse convertToResponse(Product product){
         return ProductResponse.of(product);
+    }
+
+    private boolean hasPermisson (Long sellerId) {
+        return sellerId.equals(rq.getMember().getId());
     }
 }
